Unifying camera parameter conventions in @ImageGrabber@ and @OpenNIGrabber@
git-svn-id: svn+ssh://svn.pointclouds.org/pcl/trunk@8231 a9d63959-f2ad-4865-b262-bf0e56cfafb6

+++ visualization/tools/image_grabber_viewer.cpp
@@ -202,7 +202,7 @@ main (int argc, char** argv)
   grabber->setDepthImageUnits(1E-3);
 
   // Before manually setting
-  float fx, fy, cx, cy;
+  double fx, fy, cx, cy;
   grabber->getCameraIntrinsics (fx, fy, cx, cy);
   PCL_INFO ("Factory default intrinsics: %f, %f, %f, %f\n", 
       fx, fy, cx, cy);
